search action
setup search action. this is not complete but looking for feedback.

currently it is setup to work for "search amazon for books" this will launch and browser and display an amazon search for books. you can do this for walmart, youtube, google, bing, duckduckgo, and  amazon.

i also am unsure how to get the default setting for what the default search that was selected.