search action (#39)
* search action

setup search action. this is not complete but looking for feedback.

currently it is setup to work for "search amazon for books" this will launch and browser and display an amazon search for books. you can do this for walmart, youtube, google, bing, duckduckgo, and  amazon.

i also am unsure how to get the default setting for what the default search that was selected.

* updated website search action

I believe I was able to incorporate everyting making it be able to correctly run if someone puts their specific search at the end of the phrase, instead of before the for phrase. i'm not sure how to do that right now, but I can attempt it later tonight.

* Update StoredProcedures.cs

fixed merge issue

* updated

searchParameters, foreach loop, fromdatarow. please let me know if i missed anything

* Update WebsiteSearchAction.cs

Co-authored-by: ploiu <43047560+ploiu@users.noreply.github.com>