custom responses
I came up with a greeting, you are welcomed and an age response.

Something that I noticed is if you have something that can be one or two things, it will pick what is setup first. for example, if you are searching for something that has the word time in it, if you setup the time branch first, it will ignore that it is  a search.

I also noticed since we are using a keyword, that if you say yellow yellow what is age or something goofy, it will give you your age, so I tried to narrow it by checking specific phrases. I do not know what the correct answer is, because that limits what the user can say and makes much more manual checking.